refactor(changelog): move source under github folder
