+++ lib/workers/repository/update/pr/changelog/github/source.ts
@@ -1,10 +1,10 @@
 import URL from 'node:url';
-import { GlobalConfig } from '../../../../../config/global';
-import { logger } from '../../../../../logger';
-import * as hostRules from '../../../../../util/host-rules';
-import type { BranchUpgradeConfig } from '../../../../types';
-import { ChangeLogSource } from './source';
-import type { ChangeLogError } from './types';
+import { GlobalConfig } from '../../../../../../config/global';
+import { logger } from '../../../../../../logger';
+import * as hostRules from '../../../../../../util/host-rules';
+import type { BranchUpgradeConfig } from '../../../../../types';
+import { ChangeLogSource } from '../source';
+import type { ChangeLogError } from '../types';
 export class GitHubChangeLogSource extends ChangeLogSource {
   constructor() {
     super('github', 'github-tags');
